+++ src/Command.hs
@@ -3,6 +3,7 @@ module Command (
 ) where
 
 import Value (Value)
+import Interpolation (Interpolation)
 import qualified BuiltinFunction as BF
 import qualified BuiltinModifier as BM
 import qualified BuiltinStackOp as BSO
@@ -10,12 +11,14 @@ import qualified BuiltinStackOp as BSO
 -- A Command represents the change to the program state caused by a single
 -- token of the program
 --  PushFn: Push a built-in Function to the Stack
+--  PushInterpolation: Push an Interpolation function to the Stack
 --  ModifyFn: Modify the top Function(s) on the Stack
 --  StackCmd: Modify the Stack in some way
 --  BindVal: Bind a Value to the topmost Function on the Stack
 --  BindArg: Bind one of the argument values to the topmost Function
 data Command =
   PushFn BF.BuiltinFunction |
+  PushInterpolation Interpolation |
   ModifyFn BM.BuiltinModifier |
   StackCmd BSO.BuiltinStackOp |
   BindVal Value |